+++ packages/tool.release/package.json
@@ -33,8 +33,9 @@
     "@changesets/pre": "^1.0.9",
     "@changesets/read": "^0.5.3",
     "@manypkg/get-packages": "^1.1.3",
-    "@octokit/": "^6.1.2",
-    "@octokit/plugin-throttling": "^9.1.0",
+    "octokit": "^3.2.0",
+    "@octokit/plugin-throttling": "^8.0.0",
+    "@octokit/core": "^5.2.0",
     "fs-extra": "^8.1.0",
     "mdast-util-to-string": "^1.0.6",
     "remark-parse": "^7.0.1",
